+++ crates/voicevox_core/src/engine.rs
@@ -1,21 +1,5 @@
-//! テキスト関係やAudioQuery周り、またWAV出力に関する「エンジン」の領域。
-
-mod acoustic_feature_extractor;
 mod audio_file;
-mod full_context_label;
-mod interpret_query;
-mod kana_parser;
-mod model;
-mod mora_list;
-pub(crate) mod open_jtalk;
-pub(crate) mod text_analyzer;
-pub(crate) mod user_dict;
+pub(crate) mod talk;
 
-pub(crate) use self::acoustic_feature_extractor::OjtPhoneme;
 pub(crate) use self::audio_file::to_s16le_pcm;
 pub use self::audio_file::wav_from_s16le;
-pub(crate) use self::full_context_label::extract_full_context_label;
-pub(crate) use self::interpret_query::{initial_process, split_mora, DecoderFeature};
-pub(crate) use self::kana_parser::{create_kana, parse_kana, KanaParseError};
-pub use self::model::{AccentPhrase, AudioQuery, Mora};
-pub(crate) use self::mora_list::mora2text;
