+++ crates/voicevox_core/src/engine/talk.rs
@@ -1,0 +1,18 @@
+//! テキスト関係やAudioQuery周り、またWAV出力に関する「エンジン」の領域。
+
+mod acoustic_feature_extractor;
+mod full_context_label;
+mod interpret_query;
+mod kana_parser;
+mod model;
+mod mora_list;
+pub(crate) mod open_jtalk;
+pub(crate) mod text_analyzer;
+pub(crate) mod user_dict;
+
+pub(crate) use self::acoustic_feature_extractor::OjtPhoneme;
+pub(crate) use self::full_context_label::extract_full_context_label;
+pub(crate) use self::interpret_query::{initial_process, split_mora, DecoderFeature};
+pub(crate) use self::kana_parser::{create_kana, parse_kana, KanaParseError};
+pub use self::model::{AccentPhrase, AudioQuery, Mora};
+pub(crate) use self::mora_list::mora2text;
