+++ crates/voicevox_core/src/engine/talk/open_jtalk.rs
