+++ crates/voicevox_core/src/engine/talk/user_dict/part_of_speech_data.rs
