+++ crates/voicevox_core/src/lib.rs
@@ -277,7 +277,7 @@ pub use self::{
         metas::{CharacterMeta, CharacterVersion, StyleId, StyleMeta, StyleType, VoiceModelMeta},
         voice_model::VoiceModelId,
     },
-    engine::{
+    engine::talk::{
         user_dict::{UserDictWord, UserDictWordBuilder, UserDictWordType},
         AccentPhrase, AudioQuery, Mora,
     },
